Fixed numerous places where NetMessage.m_senderEndPoint wasn't set properly. ReadMessage() overrides which outputs an IPEndPoint should be much more reliable now.

+++ trunk/Lidgren.Network/NetBuffer.Write.cs
@@ -417,7 +417,7 @@ namespace Lidgren.Network
 		public void Write(IPEndPoint endPoint)
 		{
 			Write(BitConverter.ToUInt32(endPoint.Address.GetAddressBytes(), 0));
-			Write(endPoint.Port);
+			Write((ushort)endPoint.Port);
 		}
 
 		/// <summary>
